+++ web-auth/models/role_model.py
@@ -1,9 +1,9 @@
 from .base_model import BaseModel
-from sqlalchemy import Integer, String
+from sqlalchemy import BigInteger, String
 from sqlalchemy.orm import mapped_column, Mapped
 
 
 class RoleModel(BaseModel):
     __tablename__='ROLE'
-    _id:Mapped[int]=mapped_column(Integer, primary_key=True, autoincrement=True)
+    _id:Mapped[int]=mapped_column(BigInteger, primary_key=True, autoincrement=True)
     role:Mapped[str]=mapped_column(String(35), nullable=False, unique=True)
